Use pesy as a dev dependency
See https://github.com/jchavarri/rebez/issues/4

"This is probably due to stale dune files: pesy must have marked dune files as stale after it updated. But the problem can be avoided by using pesy a bit differently.

Pesy is right now caught in between being a build tool and a config manager. It doesn't function to it full potential as a build tool as it doesn't work closely with dune enough (for instance it is forced to create dune files in source). So it's best left to just manage config for you.

This means, we now use pesy only to generated config files for the project (dune and CI). We can now move pesy to devDependencies, thereby ensuring packages that consume rebez don't have to pull pesy."