Adds support for npm's require syntax
Pesy used to assign upper camel cased directory name as the namespace
to a directory. This caused namespace clashes and can sometimes be a
source of confusion.

This commit adds support for require syntax from npm.

"imports": [
  "Foo = require("foo/library")
  "Bar = require("bar/library")
]

Earlier this would clash.

Pesy works by creating modules aliases for both the deps in a
pesy_modules directory