Stop restricting compiler version (#86)
* Stop restricting compiler version

* Update package.json

* Support 4.4 - 4.8 and drop seq from the vendored list

When dependencies: "ocaml": "4.4.2004 - 4.8" and devDependencies: "ocaml": "4.6.10 - 4.8",

Conflicting constraints:
@pesy/esy-pesy -> @reason-native/pastel -> @opam/re -> @opam/seq ->
ocaml@<4.7.0
@pesy/esy-pesy -> ocaml@=4.8.1000

ie if consuming packages want to use ocaml > 4.7, latest seq
package (which is basically empty since the compiler supports
natively) must be used. If consuming package is < 4.7, c-cube
fallback package must be used. Esy's resolutions field cannot be
used to vendor seq package - resolutions package can only be
used to pin one package at a time.

* e2e for ocaml packages (npm) 4.4 and 4.8

* Raise the lowest compiler supported to 4.6

This is because for some reason, possible due to
https://github.com/esy/esy/issues/1013, num.1.3 would get pulled

Also, num must not be vendored as pre 4.6 num comes bundled with the
compiler and must not be installed separately (do so fails)
Vendoring muliple version is not possible

* Fix: 4.4 -> 4.6

* Remove unused vendored deps

* Adds notes for compiler support

* Adds compiler support section to README

* Fix formatting